Separate creating TSDB's Scan into separate class
